[WIP] API integration, data manipulation and Movie card UI display
API integrated and the data is manipulated to suit the requirements.

Fetching from two apis to get the data. First API to get the names of
the movies and second data to get the details of the movie by ID.
This is because the first API only gives the names of the movies which
does not have all the details such as ratings or dates which can be used
for filtering.

Displaying the movie details correctly in the movie cards component.